Release/actions

+++ app/Http/Controllers/Web/System/FormController.php
@@ -45,10 +45,14 @@ class FormController extends Controller
 
     public function storeQuestion(StoreQuestionRequest $request, $uuid){
         try {
+            if(!$this->verify($uuid, 2)){
+                return redirect()->back()->with("error", "Não é possível mais adicionar perguntas pois o formulário já foi publicado.")->withInput();
+            }
+
             $this->formsRepository->createQuestion($request, $uuid);
-            return redirect()->back()->with("success", "Formulário cadastrado com sucesso.");
+            return redirect()->back()->with("success", "Pergunta cadastrada com sucesso.");
         } catch (\Throwable $th) {
-            return redirect()->back()->with("error", "Erro ao cadastrar formulário. Por favor, tente novamente mais tarde.")->withInput();
+            return redirect()->back()->with("error", "Erro ao cadastrar pergunta. Por favor, tente novamente mais tarde.")->withInput();
         }
     }
     
@@ -63,6 +67,10 @@ class FormController extends Controller
 
     public function sessionUpdate(UpdateSessionRequest $request, $uuid){
         try {
+            if(!$this->verify($uuid, 2)){
+                return redirect()->back()->with("error", "Não é possível mais editar seções pois o formulário já foi publicado.")->withInput();
+            }
+
             $this->formsRepository->updateSession($request, $uuid);
             return redirect()->back()->with("success", "Seção atualizado com sucesso.");
         } catch (\Throwable $th) {
@@ -78,6 +86,10 @@ class FormController extends Controller
 
     public function sessionDelete($uuid){
         try {
+            if(!$this->verify($uuid, 2)){
+                return redirect()->back()->with("error", "Não é possível mais deletar seções pois o formulário já foi publicado.")->withInput();
+            }
+
             $session = $this->formsRepository->getSessionById($uuid);
             $form = $this->formsRepository->getFormById($session->id_formulario);
             $this->formsRepository->deleteSessions($uuid);
@@ -95,6 +107,10 @@ class FormController extends Controller
 
     public function deleteQuestion($uuid){
         try {
+            if(!$this->verify($uuid, 3)){
+                return redirect()->back()->with("error", "Não é possível mais deletar perguntas pois o formulário já foi publicado.")->withInput();
+            }
+
             $this->formsRepository->deleteQuestion($uuid);
 
             return redirect()->back()->with("success", "Pergunta deletada com sucesso.");
@@ -105,6 +121,10 @@ class FormController extends Controller
 
     public function sessionStore(StoreSessionRequest $request, $uuid){
         try {
+            if(!$this->verify($uuid, 1)){
+                return redirect()->back()->with("error", "Não é possível mais adicionar seções pois o formulário já foi publicado.")->withInput();
+            }
+
             $this->formsRepository->storeSessions($request, $uuid);
             return redirect()->back()->with("success", "Seção adicionada com sucesso.");
         } catch (\Throwable $th) {
@@ -114,10 +134,46 @@ class FormController extends Controller
 
     public function destroy(Request $request, $uuid) {
         try {
+            if(!$this->verify($uuid, 1)){
+                return redirect()->back()->with("error", "Não é possível mais deletar pois o formulário já foi publicado.")->withInput();
+            }
+
             $this->formsRepository->destroy($uuid);
             return to_route('forms.index')->with('success', 'Formulário deletado com sucesso.');
         } catch (\Throwable $err) {
             return redirect()->back()->with('error', 'Erro ao deletar formulário. Por favor, tente novamente mais tarde.');
+        }
+    }
+
+    public function verify($uuid, $type) {
+        try {
+            if($type == 1){
+                $form = $this->formsRepository->getFormById($uuid);
+    
+                if($form->published == 1){
+                    return false;
+                }
+            }
+            if($type == 2){
+                $form_uuid = $this->formsRepository->getSessionById($uuid)->id_formulario;
+                $form = $this->formsRepository->getFormById($form_uuid);
+    
+                if($form->published == 1){
+                    return false;
+                }
+            }
+            if($type == 3){
+                $session_uuid = $this->formsRepository->getQuestionById($uuid)->id_secao;
+                $form_uuid = $this->formsRepository->getSessionById($session_uuid)->id_formulario;
+                $form = $this->formsRepository->getFormById($form_uuid);
+    
+                if($form->published == 1){
+                    return false;
+                }
+            }
+            return true;
+        } catch (\Throwable $err) {
+            return true;
         }
     }
 }
